added 'phpcassa_' prefix to class names
