+++ lib/phpcassa/Connection/MaxRetriesException.php
@@ -1,0 +1,7 @@
+<?php
+
+/**
+ * @package phpcassa
+ * @subpackage connection
+ */
+class phpcassa_Connection_MaxRetriesException extends Exception { }
