+++ lib/phpcassa/Connection/NoServerAvailable.php
@@ -1,0 +1,6 @@
+<?php
+/**
+ * @package phpcassa
+ * @subpackage connection
+ */
+class phpcassa_Connection_NoServerAvailable extends Exception { }
