+++ lib/phpcassa/Util/UUIDException.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * @package phpcassa
+ * @subpackage uuid
+ */
+class phpcassa_Util_UUIDException extends Exception {
+}
