+++ lib/thrift/autoload.php
