+++ lib/thrift/packages/cassandra/cassandra_constants.php
