+++ lib/thrift/protocol/TBinaryProtocol.php
