+++ lib/thrift/protocol/TBinarySerializer.php
