+++ lib/thrift/transport/TBufferedTransport.php
