+++ lib/thrift/transport/TFramedTransport.php
