+++ lib/thrift/transport/THttpClient.php
