+++ lib/thrift/transport/TMemoryBuffer.php
