+++ lib/thrift/transport/TNullTransport.php
