+++ lib/thrift/transport/TPhpStream.php
