+++ lib/thrift/transport/TServerTransport.php
