+++ lib/thrift/transport/TTransport.php
