+++ test/test_columnfamily.php
@@ -1,7 +1,6 @@
 <?php
 require_once('simpletest/autorun.php');
-require_once('../connection.php');
-require_once('../columnfamily.php');
+require_once('../lib/autoload.php');
 
 class TestColumnFamily extends UnitTestCase {
 
@@ -11,8 +10,8 @@ class TestColumnFamily extends UnitTestCase {
     private static $KEYS = array('key1', 'key2', 'key3');
 
     public function setUp() {
-        $this->pool = new Connection('Keyspace1');
-        $this->cf = new ColumnFamily($this->pool, 'Standard1');
+        $this->pool = new phpcassa_Connection('Keyspace1');
+        $this->cf = new phpcassa_ColumnFamily($this->pool, 'Standard1');
     }
 
     public function tearDown() {
@@ -110,7 +109,7 @@ class TestColumnFamily extends UnitTestCase {
 
     public function test_insert_get_range() {
         $cl = cassandra_ConsistencyLevel::ONE;
-        $cf = new ColumnFamily($this->pool,
+        $cf = new phpcassa_ColumnFamily($this->pool,
                                'Standard1', true, true,
                                $read_consistency_level=$cl,
                                $write_consistency_level=$cl,
@@ -138,7 +137,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Buffer size larger than row count
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=1000);
         $count = 0;
@@ -151,7 +150,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Buffer size larger than row count, less than total number of rows
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=150);
         $count = 0;
@@ -164,7 +163,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Odd number for batch size
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=7);
         $count = 0;
@@ -177,7 +176,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Smallest buffer size available
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=2);
         $count = 0;
@@ -195,7 +194,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Row count above total number of rows
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=2);
         $count = 0;
@@ -208,7 +207,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Row count above total number of rows
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=7);
         $count = 0;
@@ -222,7 +221,7 @@ class TestColumnFamily extends UnitTestCase {
 
  
         # Row count above total number of rows, buffer_size = total number of rows
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=200);
         $count = 0;
@@ -235,7 +234,7 @@ class TestColumnFamily extends UnitTestCase {
  
  
         # Row count above total number of rows, buffer_size = total number of rows
-        $cf = new ColumnFamily($this->pool, 'Standard1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Standard1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=10000);
         $count = 0;
@@ -253,7 +252,7 @@ class TestColumnFamily extends UnitTestCase {
     public function test_batched_get_indexed_slices() {
 
         $cl = cassandra_ConsistencyLevel::ONE;
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=10);
         $cf->truncate();
@@ -269,9 +268,8 @@ class TestColumnFamily extends UnitTestCase {
         foreach (range(201, 300) as $i)
             $cf->insert('key'.$i, $columns);
 
-
-        $expr = CassandraUtil::create_index_expression($column_name='birthdate', $value=1);
-        $clause = CassandraUtil::create_index_clause(array($expr), 100);
+        $expr = phpcassa_Util_CassandraUtil::create_index_expression($column_name='birthdate', $value=1);
+        $clause = phpcassa_Util_CassandraUtil::create_index_clause(array($expr), 100);
 
         # Buffer size = 10; rowcount is divisible by buffer size
         $count = 0;
@@ -283,7 +281,7 @@ class TestColumnFamily extends UnitTestCase {
         self::assertEqual($count, 100);
 
         # Buffer size larger than row count
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=1000);
         $count = 0;
@@ -296,7 +294,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Buffer size larger than row count, less than total number of rows
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=150);
         $count = 0;
@@ -309,7 +307,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Odd number for batch size
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=7);
         $count = 0;
@@ -322,7 +320,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Smallest buffer size available
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=2);
         $count = 0;
@@ -341,7 +339,7 @@ class TestColumnFamily extends UnitTestCase {
 
         # Row count above total number of rows
         $clause->count = 10000;
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=2);
         $count = 0;
@@ -354,7 +352,7 @@ class TestColumnFamily extends UnitTestCase {
 
 
         # Row count above total number of rows
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=7);
         $count = 0;
@@ -368,7 +366,7 @@ class TestColumnFamily extends UnitTestCase {
 
  
         # Row count above total number of rows, buffer_size = total number of rows
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=200);
         $count = 0;
@@ -381,7 +379,7 @@ class TestColumnFamily extends UnitTestCase {
  
  
         # Row count above total number of rows, buffer_size = total number of rows
-        $cf = new ColumnFamily($this->pool, 'Indexed1', true, true,
+        $cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1', true, true,
                                $read_consistency_level=$cl, $write_consistency_level=$cl,
                                $buffer_size=10000);
         $count = 0;
@@ -396,7 +394,7 @@ class TestColumnFamily extends UnitTestCase {
     }
 
     public function test_get_indexed_slices() {
-        $indexed_cf = new ColumnFamily($this->pool, 'Indexed1');
+        $indexed_cf = new phpcassa_ColumnFamily($this->pool, 'Indexed1');
         $indexed_cf->truncate();
 
         $columns = array('birthdate' => 1);
@@ -404,8 +402,8 @@ class TestColumnFamily extends UnitTestCase {
         foreach(range(1,3) as $i)
             $indexed_cf->insert('key'.$i, $columns);
 
-        $expr = CassandraUtil::create_index_expression($column_name='birthdate', $value=1);
-        $clause = CassandraUtil::create_index_clause(array($expr), 10000);
+        $expr = phpcassa_Util_CassandraUtil::create_index_expression($column_name='birthdate', $value=1);
+        $clause = phpcassa_Util_CassandraUtil::create_index_clause(array($expr), 10000);
         $result = $indexed_cf->get_indexed_slices($clause);
 
         $count = 0;
@@ -483,8 +481,8 @@ class TestSuperColumnFamily extends UnitTestCase {
     private static $KEYS = array('key1', 'key2', 'key3');
 
     public function setUp() {
-        $this->pool = new Connection('Keyspace1');
-        $this->cf = new ColumnFamily($this->pool, 'Super1');
+        $this->pool = new phpcassa_Connection('Keyspace1');
+        $this->cf = new phpcassa_ColumnFamily($this->pool, 'Super1');
     }
 
     public function tearDown() {
